requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0
jupyter>=1.0.0
notebook>=7.0.0
gradio>=4.0.0